chore(tsify): upgrade `tsify` version to 1.0.1
The current version of browserify-typescript is using tsify@^0.14.1
which is incompatible with @types from TypeScript 2.0. A fix has been
created in tsify@^1.0.1.
By upgrading the package.json with tsify@^1.0.1, TypeScript becomes a
devDependency and it is not installed by tsify anymore. TypeScript will
have to be installed via npm install typescript --save-dev in your
project to use whatever version you want.

fix #42